requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
mysql-connector==2.2.9
numpy==1.19.5
Pillow==8.0.1
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
